finished client controller

+++ src/main/java/com/tienda/persistent/entity/Cliente.java
@@ -13,7 +13,7 @@ public class Cliente {
 
     private String nombre;
     private String apellidos;
-    private Integer celular;
+    private Double celular;
     private String direccion;
 
     @Column( name = "correo_electronico")
@@ -46,11 +46,11 @@ public class Cliente {
         this.apellidos = apellidos;
     }
 
-    public Integer getCelular() {
+    public Double getCelular() {
         return celular;
     }
 
-    public void setCelular(Integer celular) {
+    public void setCelular(Double celular) {
         this.celular = celular;
     }
